Move common function into existing file
